requirements.txt
langchain
langchain-community
python-dotenv
sentence_transformers
django
django-tailwind[reload]
PyMySQL
pinecone-client
langchain_groq
Markdown
djangorestframework
django-jazzmin
einops
gunicorn
pypdf
